treaty: include bill and seal alongside docket
In addition to serving the docket file over the network, we now also
include the bill and seal files for a desk. It is important for other
ships to be able to inspect these prior to attempting an install of the
desk: they will need to grant the permissions from the seal file for it
to be able to run at all.

The bill file helps clarify some of the agent->app mappings that may be
needed for rendering the permissions, and also lets clients check for
agent conflicts ahead of time.

Some unfinished work herein. There is no state adapter. There is no
backwards compatibility (outside of the bare minimal logic).