treaty: treaty subscription back-compat
The previous commit added some versioned treaty subscription
affordances, but didn't actually do the required legwork of version
negotiation to make this useful.

Here, we implement a /version subscription endpoint into treaty, and use
it for version negotiation. A nack means they're still on v0. An ack
is expected to be followed by a fact containing an indication of v1, in
response to which we leave all outgoing old-style subscriptions and
replace them with new-style ones.

To try and upgrade all existing subscriptions, we make sure to retry
version negotiation whenever an old-style subscription gets kicked. When
treaty upgrades, it kicks all incoming old-style subscriptions, and
attempts version negotiation with everyone it's subscribed to.

We track versions we've negotiated in state. Before we send new
subscriptions, we want to know the version, and so track "pending treaty
subscriptions" in state also.

Incidentially we also remove :direct from state, which was getting
properly book-kept, but not actually used for anything.