treaty: fix risk of "bunted bowl" type bugs
In +on-peek, we were using +ca instead of +cc to access +cg.
This should work for the exact things we use there, but will silently
break as soon as we start using any data from the bowl in the code being
invoked. So, we simply replace these invocations with :cc instead.