notifications: fix duplicate notification issue
I wasn't able to reproduce the exact issue in #125, but I think this was actually what the user was seeing.

We were firing off a test notification each time the count or unreadNotifications changed, along with the real notification itself.

We also weren't properly including notification content (DMs or mentions would just show that you got a DM or a mention, but not the content of the message). This fixes that.