state: handling bad vase appropriately
We saw bad behavior upon merging the prerelease where the agents were
already present but archived. This seems to be the result of weird
syncing or some other dev ops. This resulted in the old state being the bunt of vase (an atom of 0).

Here we catch that case and call `on-init` to re-initialize the state.