lib: avoid crash during activity summarization
There is a (somewhat poorly understood) case where the group associated
with a chat may not actually contain that chat in its channels listing.

Here, instead of assuming the channel is always present, we ask for it
more gently, and simply skip to the next-most-active channel if it turns
out to be missing.

We also remove a stray debug printf, and add a trace print in case
anything else goes wrong here.