bark: don't crash on no-mail summary saving
Instead, give a %no-mail response message. It's not entirely unexpected
to encounter ships that don't have a hosting account email address
associated with them, so it shouldn't be a reason to crash.

The crash is particularly problematic because as of the previous
commits, we now actually show the trace of thread crashes.