treaty: update treaty version
this exposes a gall bug that triggers a kick-rewatch loop on the
subscriber sides. gall does not register the mark conversion failure
from treaty-0 to treaty-1 as an error and does not send a watch-nack, so
the subscriber will keep trying to watch the publisher's treaty even
though it immediately gets kicked.