growl: don't fumble the activity logging consent
Previous logic for reading %settings-event facts was a little to eager
in declaring explicit non-consent. It would also set the flag to false
in cases where we received updates not matching what we were looking
for.

Now, it's not entirely certain that this was a problem, but we did
observe "disabled growl" on ships that had the setting toggle on the
frontend set to true, and this seemed like the only viable cause of
that.

We also update +on-load to re-set the flag from the current settings
value, to correct any lingering mismatching states.