growl: don't scry during +on-load
The upgrade from %0 to %1 was calling into +on-init, which issues a
scry. In some cases, that might not be able to complete successfully,
aborting the whole event that triggered the load.

Here, we move this "set flag based on settings" logic into a %default
poke command, and change the +on-load logic to set a timer that calls
that instead, defering the scry to the next event.