docket: suspend the %garden desk
It was deprecated in favor of %landscape, but is still active on
people's ships and blocking the 411 update. Here, we make docket suspend
the %garden desk unconditionally.

Debated deferring this to a separate event, since it may kick off the
base update, but the suspension itself should isolate that into its own
separate event already.